requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
yfinance==0.2.28
pandas==2.1.3
python-dotenv==1.0.0
tweepy==4.14.0
requests==2.31.0
transformers==4.35.2
torch==2.1.1
google-generativeai==0.3.2
numpy==1.24.3
scikit-learn==1.3.2